feat: Update logo and navigation links in theme configuration
This commit updates the logo and navigation links in the theme configuration of the Vitepress site. The logo is now a clickable image with a link to "https://www.penify.dev". The "Home" link has been updated to point to "https://www.penify.dev" and a new "Main" link has been added to navigate to the root of the site. These changes enhance the user experience by providing consistent branding and improved navigation options.